+++ src/infrastructure/resources/git.py
@@ -23,9 +23,7 @@ from .acceleration import (
 from .manifest import ResourceManifest
 from .paths import default_resource_repository_dir, resource_repository_dir
 
-DEFAULT_RESOURCE_REMOTE = (
-    "https://github.com/FlanChanXwO/astrbot_plugin_dna_resources.git"
-)
+DEFAULT_RESOURCE_REMOTE = "https://github.com/FlanChanXwO/dna-resource.git"
 
 
 class ResourceSyncError(RuntimeError):